Computer Vision Problem Statement
Multi-Class Image Classification
Problem Details:
You have to perform Multi-Class Image Classfication on the given data.
There are 10 classes in the dataset.
Download the dataset from here.
To know more about the dataset, check out this link.
Bonus:
Perform image detection and recognition from a live url.
The script in available in the folder. "face_detector_live_video.py".
Or you can develop frame from the video and run the model on it.